Moved specs for the JDBC view in separate folder
